Perform null check before updating originating tokens
Turns out it's possible that when the callback is processed there is no
process attached to the ServiceRecord despite it being allowed for bg
activity starts "by start". This can happen for example if we call
allowBgActivityStartsOnServiceStart() twice without a process attached.

This was handled before ag/12559801, so also needs to be handled in the
new code-path.

Bug: 176813978
Test: atest BackgroundActivityLaunchTest
Change-Id: I2024f5cb4da78a93c868a63e45c00567c20848ba