Fix app in size compat when it should not
1. CompatDisplayInsets now stores the Task bounds instead of RootDA
bounds so that the activity will not be resized regardless the windowing
mode it is currently in.
2. Because the compat app container may not fill the display, update the
insets to only contain the part that intersect with the container.
3. Before when orientation is ignored and app enters task letterbox, it
shows as size compat after rotate 180/360 degree, because task won't go
letterbox again when compat exists. Now, it will enter task letterbox if
the compat is also from task letterbox of the same size.

Bug: 173124775
Bug: 176061101
Test: atest WmTests:SizeCompatTests
Change-Id: I319ccb3c357e467356e23a06b09cefb80ab7d536