Revert^2 "Set setCurrentProxyScriptUrl as public"
PacManager is part of services.jar, while ProxyManager is in
service-connectivity.jar. setCurrentProxyScriptUrl cannot be
package-private and called across jars.

As part of modularization of service-connectivity.jar, the
PacManager interface will then be formally defined.

This rolls forward the change. The original topic was reverted because
of a bad merged-in clause; this is fixed and re-verified in this topic.

Bug: 171540887
Test: Splitting service-connectivity, ProxyManager can reference the
      method

Change-Id: I1007be6c0101c4582bade182f9bc8dcfc4a1b303