Force update IME container relative layer when IME parent is changed
Before, WindowContainer#assignRelativeLayer() will not set layer if the
call is the same as the previous call. This could cause issue.

In case the previous IME parent is removed, the relative layer of IME is
also removed. However, WindowContainer won't know about that, so the
next assignRelativeLayer with the same arguments will still be ignored.

Now we do a force update when the IME parent is changed.

Bug: 180541134
Test: atest WmTests:DisplayContentTests
Test: atest WmTests:WindowContainerTests
Test: verify with reproduce steps in bug
Change-Id: Ie1f00146f150a5fcf818d775239bd93d5235ead3