Return copy of pending token list
Make SyntheticPasswordManager.getPendingTokensForUser() return
a copy of the token list, as the underlying list might change
if a pending token is activated or removed. This fixes the NPE
in LockSettingsService.activateEscrowTokens() where it tries
to iterate through the list and activate tokens.

Bug: 178560581
Test: atest com.android.server.locksettings

Merged-In: I59d159d792a3d9e2a79ea521d6f01a5cb7c65695
Change-Id: I59d159d792a3d9e2a79ea521d6f01a5cb7c65695