Move text to new line in big base style
This was only used when progress was shown, did not exist for headerless, and was then hidden for all styles; this just left the big base template, where it needed more room.  Now we just only set the text with progress when using that template.

Note: The line1 container is used for margin adjustments, so removing it is non-trivial; I'll do that in a followup CL.

Bug: 181048615
Test: visual inspection
Change-Id: I465f7be95249b90ae8906b2bc951c9ff6066550a