DMD: Also consider no votes when calculating specs
Currently when DisplayModeDirector is calculating desired
display mode specs, we consider different ranges of vote
priorities. If we can't find allowed modes with the broadest
range we increase the lowest priority. However we always
take into account the vote with highes priority. It is
possible that even when we consider only that vote, we
end up with no allowed modes. This currently happens only in
tests when AlwaysRespectAppRequest is ON.

This CL changes the failure behaviour to from throwing ISE
to falling back to the default mode.

Bug: 182387031
Test: atest DisplayModeDirectorTest
Change-Id: I708d6bce0ceddd3c439e66eecd5193f82e272109